Use dynamic names for the name of the containers
This commit modifies the docker-compose configuration file to use dynamic
names instead of hardcoded names for the MySQL and WP containers. The
goal is to  make it easier to preserve changes when
switching from one trial candidate dev environment to another. Before
this, since the containers for different candidates shared the same name,
it was necessary to remove the containers, and thus lose all the
changes when switching from one environment to another.

Without the hardcoded names, docker-compose will create a different name,
based on the name of the root directory of the repository, for each
container of each trial candidate environment.

Two scripts (bin/setup.sh and bin/reset-fixtures.sh) had hardcoded the
name of the WP container. I changed those to get the name of the
container dynamically using the name of the service.