Pistachio Admin Page UI/UX Improvements (#93)
Applied a lot of default WP styles to the table. A major change is that the table overflow is hidden within a container so you don't see it all until scrolling. We could update to use breakpoints to rearrange layout at a later time.

* Update actions list design

I tried to find a place where the actions list would meet a few pieces of feedback we've heard:
- It should look more modern (like the three dot dropdowns)
- Part of the point of showing/hiding things is to not take up as much space. The previous design had the full space for everything, but hid it. This removes that extra space.
- A dropdown would normally go underneath, but the dropdown + hover showing lots of them made them overlap and got overwhelming quickly. I moved it to the left and covered up some of the existing row data so it would be mostly the same interaction for ease of testing.

* Added aria-label to show/hide actions button