Change candidates table structure to focus on most useful info (#98)
The candidates table had a few unused columns (needs action and last activity) that were cluttering the view. Those were removed, and all `needs_action` code was removed as well (which was a lot more than expected). The last activity data structure remains since it may be used in the future, but would need improvement before being added back to the table. 

A secondary improvement was rearranging output so the region and pronoun information is further to the left. The order of the info from left to right is more candidate personal info first, and then moving to coordinator and candidate upload info on the right.